backend/claude/requirements.txt
requests
python-dotenv
fastapi
uvicorn
anthropic
openai